// stream/enum/event-name.enum.ts
export enum EventNameEnum {
  ON_USER_DISCONNECTED_FROM_TALK = "ON_USER_DISCONNECTED_FROM_TALK",
  ON_UPDATE_CONSUMER = "ON_UPDATE_CONSUMER",
  ON_UPDATE_VIDEO_STREAM = "ON_UPDATE_VIDEO_STREAM",
  ON_FINISH_BROADCAST = "ON_FINISH_BROADCAST",
  ON_NEED_STREAM_ACCESS = "ON_NEED_STREAM_ACCESS",
  ON_NEED_STREAM_TABS = "ON_NEED_STREAM_TABS",
  ON_UPDATE_STREAM_CURRENT_USER = "ON_UPDATE_STREAM_CURRENT_USER",
  ON_UPDATE_STREAM_HAS_TALK_USERS = "ON_UPDATE_STREAM_HAS_TALK_USERS",
  ON_UPDATE_AMA_SPEAKERS = "ON_UPDATE_AMA_SPEAKERS",
  ON_GET_NEW_STREAM_MESSAGE = "ON_GET_NEW_STREAM_MESSAGE",
  ON_CLOSE_STREAM_TABS_DIALOG = "ON_CLOSE_STREAM_TABS_DIALOG",
  ON_SHOW_TOAST = "ON_SHOW_TOAST",
}